requirements.txt
streamlit==1.36.0
pandas==2.2.2
openpyxl==3.1.2
reportlab==4.2.0
Pillow==10.4.0
requests==2.32.3
